Make nodes_from_layer a funtion, not a method
